fix(build): remove all duplicate type definitions, fix type conversions
- Remove duplicate ServerNode, DNSConfiguration, RoutingRule, GeoFile,
  V2XSubscription, ThreatEvent, LogEntry from VPNProfile.swift
  (canonical definitions remain in NetworkModels.swift and URLSchemeItem.swift)
- Rename ContentView's SplashParticle to ContentSplashParticle to avoid
  conflict with SplashView's SplashParticle
- Fix Double→Int64 conversion in ShareProxyView and NetworkDiagnosticsView
- Fix Int→Double conversion in ThreatShieldView for numericText

+++ V2X/V2X/Views/NetworkDiagnosticsView.swift
@@ -419,9 +419,9 @@ struct NetworkDiagnosticsView: View {
                 sectionDivider
                 interfaceRow(label: "BSSID", value: networkMonitor.bssid)
                 sectionDivider
-                interfaceRow(label: "RX Bytes", value: String.formatBytes(networkMonitor.rxBytes))
+                interfaceRow(label: "RX Bytes", value: String.formatBytes(Int64(networkMonitor.rxBytes)))
                 sectionDivider
-                interfaceRow(label: "TX Bytes", value: String.formatBytes(networkMonitor.txBytes))
+                interfaceRow(label: "TX Bytes", value: String.formatBytes(Int64(networkMonitor.txBytes)))
                 sectionDivider
                 interfaceRow(label: "RX Packets", value: "\(networkMonitor.rxPackets)")
                 sectionDivider
